+++ client/src/App.css
@@ -38,5 +38,8 @@
   }
   to {
     transform: rotate(360deg);
-  }
+ }
 }
+
+
+
