+++ client/src/components/header.js
@@ -1,5 +1,5 @@
 import React from "react";
-import Nav from "../Nav";
+import Nav from "../../../loco-local/src/components/Nav";
 import locologo from '../../assets/icon/locologo.png'
