+++ client/src/index.css
@@ -1,13 +1,303 @@
+/* styles for components starts here
+
+:root {
+  --primary:RGB(143, 62, 63);
+  --secondary:white;
+  --tertiary:rgba(163, 127, 80, 0.815);
+  --quaternary:#adad9f;
+  --spacing-one: calc(1rem * 0.5);
+	--spacing-two: calc(1rem);
+	--spacing-five: calc(2.5rem)
+}
+main {
+	margin: 0 auto;
+	max-width: 90%;
+	padding: 50px 30px;
+}
 body {
-  margin: 0;
-  font-family: -apple-system, BlinkMacSystemFont, 'Segoe UI', 'Roboto', 'Oxygen',
-    'Ubuntu', 'Cantarell', 'Fira Sans', 'Droid Sans', 'Helvetica Neue',
-    sans-serif;
-  -webkit-font-smoothing: antialiased;
-  -moz-osx-font-smoothing: grayscale;
+	margin: 0;
+	padding: 0;
+	font-weight: 600;
+	line-height: 1.5;
+	font-size: 18px;
+	letter-spacing: .1em;
+	background: whitesmoke;
+	color: var(--primary);
+	font-family: 'Franklin Gothic Medium', 'Arial Narrow', Arial, sans-serif;
+  
+}
+a
+{
+    text-decoration: none;
+	color: var(--primary); 
+    
 }
 
-code {
-  font-family: source-code-pro, Menlo, Monaco, Consolas, 'Courier New',
-    monospace;
+h1,
+h2,
+h3{
+	font-family: 'Roboto', sans-serif;
+	font-weight: 500;
+	margin: 5px 0;
+	line-height: 1.25;
 }
+
+h1 {
+	font-size: 2.25rem;
+}
+
+h2 {
+	font-size: 2rem;
+}
+
+h3 {
+	font-size: 1.75rem;
+}
+
+p {
+	font-size: 1.1rem;
+	margin-top: 0;
+}
+
+
+
+p {
+	font-size: 1.1rem;
+	margin-top: 0;
+}
+header {
+	background: white;
+	display: flex;
+	justify-content: space-between;
+	padding: 1rem;
+	text-align: center;
+}
+
+span[role="img"] {
+	cursor: default;
+}
+
+/*nav section */
+
+ol li {
+	list-style: inside;
+}
+
+ul li {
+	list-style: none;
+}
+
+nav li {
+	cursor: pointer;
+}
+.navActive {
+	color: var(--tertiary);
+	
+}
+
+/* end nav section */
+
+img {
+	max-width: 20%;
+	float: left;
+	
+}
+.container {
+	width: 85%;
+	max-width: 1200px;
+	margin-left: auto;
+	margin-right: auto;
+}
+
+.flex-row {
+	display: flex;
+	justify-content: space-evenly;
+
+}
+
+.space-between {
+	justify-content: space-between;
+}
+
+.flex-end {
+	justify-content: flex-end;
+}
+
+.mx-1 {
+	margin-right: var(--spacing-one);
+	margin-left: var(--spacing-one);
+}
+
+.mx-2 {
+	margin-right: var(--spacing-two);
+	margin-left: var(--spacing-two);
+}
+
+.my-1 {
+	margin-top: var(--spacing-one);
+	margin-bottom: var(--spacing-one);
+}
+
+.my-2 {
+	margin-top: var(--spacing-two);
+	margin-bottom: var(--spacing-two);
+}
+
+
+.my-5 {
+	margin-top: var(--spacing-five);
+	margin-bottom: var(--spacing-five);
+}
+
+.px-1 {
+	padding-right: var(--spacing-one);
+	padding-left: var(--spacing-one);
+}
+
+.px-2 {
+	padding-right: var(--spacing-two);
+	padding-left: var(--spacing-two);
+}
+
+.py-1 {
+	padding-top: var(--spacing-one);
+	padding-bottom: var(--spacing-one);
+}
+
+.py-2 {
+	padding-top: var(--spacing-two);
+	padding-bottom: var(--spacing-two);
+}
+
+
+
+/* footer section */
+
+.footer {
+	padding:40px 0;
+	background-color:#ffffff;
+	color: RGB(143, 62, 63);
+  }
+  
+  .footer ul {
+	padding:0;
+	list-style:none;
+	text-align:center;
+	font-size:18px;
+	line-height:1.6;
+	margin-bottom:0;
+  }
+  
+  .footer li {
+	padding:0 10px;
+  }
+  
+  .footer ul a {
+	color:inherit;
+	text-decoration:none;
+	opacity:0.8;
+  }
+  
+  .footer ul a:hover {
+	opacity:1;
+  }
+  
+  .footer .social {
+	text-align:center;
+	padding-bottom:25px;
+  }
+  
+  .footer .social > a {
+	font-size:24px;
+	width:40px;
+	height:40px;
+	line-height:40px;
+	display:inline-block;
+	text-align:center;
+	border-radius:50%;
+	border:1px solid #ccc;
+	margin:0 8px;
+	color:inherit;
+	opacity:0.75;
+  }
+  
+  .footer .social > a:hover {
+	opacity:0.9;
+  }
+  
+  .footer .copyright {
+	margin-top:15px;
+	text-align:center;
+	font-size:13px;
+	color:RGB(143, 62, 63);;
+	margin-bottom:0;
+  }
+
+  /* end footer section */
+
+
+.center {
+	display: flex;
+	justify-content: center;
+}
+
+.page-header {
+	font-family: 'Birthstone Bounce', cursive;
+	font-size: 2.5em;
+	font-weight: 500;
+	margin: 5px 0;
+	line-height: 1.25;
+	padding-bottom: 2rem;
+	
+}
+
+
+
+/* Viewports Styling start*/
+@media screen and (max-width: 980px) {}
+
+@media screen and (max-width: 817px) {
+	header div {
+		position: relative;
+		width: 110%;
+	}
+
+	header div h2 {
+		position: absolute;
+		top: 50%;
+		transform: translateY(-50%);
+	}
+
+	h2 {
+		font-size: 1.5rem;
+	}
+
+	nav ul li span {
+		font-size: 0.75em;
+	}
+
+	.mobile-header {
+		width: 110%;
+	}
+
+	.mobile-row {
+		display: block;
+		margin: 0 100px 0 auto;
+	}
+
+	.mobile-view {
+		display: inline;
+		text-align: end;
+	}
+
+	
+	ol {
+		font-size: 0.75rem;
+	}
+}
+
+@media screen and (max-width: 768px) {}
+
+@media screen and (max-width: 575px) {}
+
+/* Viewports Styling end *
